+++ e2e/flows/copy-workspace-with-env.e2e.2.js
@@ -18,7 +18,7 @@ describe('copy workspace with env', function () {
       helper.setNewLocalAndRemoteScopes();
       helper.createComponentBarFoo();
       helper.addComponentBarFoo();
-      helper.importCompiler('bit.envs/compilers/react');
+      helper.importCompiler('bit.envs/compilers/react-typescript');
 
       copiedPath = helper.cloneLocalScope();
       // remove the original workspace so then symlinks get invalid
@@ -35,7 +35,7 @@ describe('copy workspace with env', function () {
       const results = parsedDoctor.examineResult.bareResult;
       expect(results.valid).to.be.false;
       expect(results.data.brokenSymlinks[0].pathToDelete).to.have.string(
-        path.normalize('.bit/components/compilers/react/bit.envs')
+        path.normalize('.bit/components/compilers/react-typescript')
       );
     });
   });