+++ CHANGELOG.md
@@ -7,10 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+## [14.1.4-dev.6] - 2019-07-02
+
+- fix require statements to an internal package file to not include extensions if they're [.js, .ts, .tsx, .jsx]
+- [#1762](https://github.com/teambit/bit/issues/1762) allow compilers to add properties to `package.json` file
 - change dependency links generated when dependencies are saved as components to be module paths and not relative paths
 - add a custom entry point file for Angular components
 - [#1750](https://github.com/teambit/bit/issues/1750) improve the output to clarify when a dependency package is missing
 - [#1752](https://github.com/teambit/bit/issues/1752) fix dependency links generation when originally there were multiple link files
+- await for a promise calls on post-add hook
 
 ## [14.1.4-dev.5] - 2019-06-25